desktop-clients/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mandala-sync"
version = "0.1.0"
description = "CLI tool for syncing local folders with Mandala Document Indexer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Mandala Team"}
]
keywords = ["document", "sync", "watcher", "upload"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "httpx>=0.24.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
    "keyring>=24.0.0",
    "rich>=13.0.0",
    "python-daemon>=3.0.0;platform_system!='Windows'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
mandala-sync = "mandala_sync.cli:main"

[project.urls]
Homepage = "https://github.com/mandala/document-indexer"
Documentation = "https://github.com/mandala/document-indexer#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["mandala_sync*"]
